implementazione "compra" in UtenteController e modifica classe ordine

+++ Models/Classes/Ordine.cs
@@ -6,7 +6,7 @@ public class Ordine : Entity
 {
     public DateTime Data { get; set; }
     public string Tipo_Ordine { get; set; }
-    public float Totale { get; set; }
+    public double Totale { get; set; }
     public string Stato { get; set; }
 
     public Utente Utente { get; set; }
